requirements.txt
click==8.1.7
pillow==11.0.0
